Excepition handling

+++ src/main/java/com/dscommerce/devsuperior/services/ProductService.java
@@ -3,10 +3,15 @@ package com.dscommerce.devsuperior.services;
 import com.dscommerce.devsuperior.dto.ProductDTO;
 import com.dscommerce.devsuperior.entities.Product;
 import com.dscommerce.devsuperior.repositories.ProductRepository;
+import com.dscommerce.devsuperior.services.exceptions.DatabaseException;
+import com.dscommerce.devsuperior.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -20,7 +25,8 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id){
-        Product product = repository.findById(id).get();
+        Product product = repository.findById(id).orElseThrow(
+                ()-> new ResourceNotFoundException("Recurso nao encontrado"));
         return new ProductDTO(product);
 
     }
@@ -42,10 +48,15 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto){
-        Product entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+        try {
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        }
+        catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso nao encontrado");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
@@ -55,9 +66,17 @@ public class ProductService {
         entity.setImgUrl(dto.getImgUrl());
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void  delete(Long id){
-        repository.deleteById(id);
+       if (!repository.existsById(id)){
+           throw new ResourceNotFoundException("recurso nao encontrado");
+       }
+       try{
+           repository.deleteById(id);
+       }
+       catch (DataIntegrityViolationException e){
+           throw new DatabaseException("falha na integridade referencial");
+       }
 
     }
